requirements-nightly.txt
# Core deps.
tf-nightly
tensorflow-text-nightly

# Common deps.
-r requirements-common.txt
